# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Little Hiker Friends API
  description: 등산 그룹 위치 공유 및 메시징 API
  version: 1.0.0
  contact:
    name: API Support
    email: support@littlehikerfriends.com

servers:
  - url: http://localhost:8080/api
    description: 로컬 개발 서버
  - url: https://api.littlehikerfriends.com/api
    description: 프로덕션 서버

tags:
  - name: Auth
    description: 인증 관련 API
  - name: User
    description: 사용자 관리 API
  - name: Hike
    description: 등산 모임 관리 API
  - name: Location
    description: 위치 관리 API
  - name: Message
    description: 메시지 관리 API

paths:
  /auth/login:
    post:
      tags:
        - Auth
      summary: 로그인
      description: 이메일과 비밀번호로 로그인합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "password123"
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    $ref: '#/components/schemas/UserResponse'
        '401':
          description: 인증 실패

  /locations:
    post:
      tags:
        - Location
      summary: 위치 업데이트
      description: 사용자의 현재 위치를 업데이트합니다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationRequest'
      responses:
        '200':
          description: 위치 업데이트 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '400':
          description: 잘못된 요청 데이터
        '401':
          description: 인증 필요

  /locations/hikes/{hikeId}/members:
    get:
      tags:
        - Location
      summary: 등산 모임 멤버들의 위치 조회
      description: 특정 등산 모임에 참여한 멤버들의 최신 위치를 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: hikeId
          in: path
          required: true
          schema:
            type: integer
            example: 1
          description: 등산 모임 ID
      responses:
        '200':
          description: 위치 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LocationResponse'
        '404':
          description: 등산 모임을 찾을 수 없음

  /locations/nearby:
    get:
      tags:
        - Location
      summary: 주변 위치 조회
      description: 특정 지점 주변의 위치들을 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: latitude
          in: query
          required: true
          schema:
            type: number
            format: double
            example: 37.5665
          description: 위도
        - name: longitude
          in: query
          required: true
          schema:
            type: number
            format: double
            example: 126.9780
          description: 경도
        - name: radiusMeters
          in: query
          required: false
          schema:
            type: number
            format: double
            default: 1000
            example: 1000
          description: 검색 반경 (미터)
      responses:
        '200':
          description: 주변 위치 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LocationResponse'

  /hikes:
    get:
      tags:
        - Hike
      summary: 등산 모임 목록 조회
      description: 등산 모임 목록을 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PLANNED, IN_PROGRESS, COMPLETED]
          description: 등산 모임 상태 필터
        - name: mountainName
          in: query
          required: false
          schema:
            type: string
          description: 산 이름으로 검색
      responses:
        '200':
          description: 등산 모임 목록 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HikeResponse'

    post:
      tags:
        - Hike
      summary: 등산 모임 생성
      description: 새로운 등산 모임을 생성합니다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HikeRequest'
      responses:
        '201':
          description: 등산 모임 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HikeResponse'

  /hikes/{hikeId}/messages:
    get:
      tags:
        - Message
      summary: 등산 모임 메시지 조회
      description: 특정 등산 모임의 메시지들을 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: hikeId
          in: path
          required: true
          schema:
            type: integer
          description: 등산 모임 ID
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: 페이지 번호
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: 페이지 크기
      responses:
        '200':
          description: 메시지 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/MessageResponse'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer

    post:
      tags:
        - Message
      summary: 메시지 전송
      description: 등산 모임에 메시지를 전송합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: hikeId
          in: path
          required: true
          schema:
            type: integer
          description: 등산 모임 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageRequest'
      responses:
        '201':
          description: 메시지 전송 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LocationRequest:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          example: 37.5665
          description: 위도
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          example: 126.9780
          description: 경도
        altitude:
          type: number
          format: double
          example: 123.5
          description: 고도 (미터)

    LocationResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        userId:
          type: integer
          example: 123
        userNickname:
          type: string
          example: "등산러버"
        latitude:
          type: number
          format: double
          example: 37.5665
        longitude:
          type: number
          format: double
          example: 126.9780
        altitude:
          type: number
          format: double
          example: 123.5
        timestamp:
          type: string
          format: date-time
          example: "2025-08-15T10:30:00"

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 123
        email:
          type: string
          format: email
          example: "user@example.com"
        nickname:
          type: string
          example: "등산러버"
        imageUrl:
          type: string
          example: "https://example.com/profile.jpg"
        provider:
          type: string
          enum: [EMAIL, GOOGLE, APPLE, KAKAO]
          example: "EMAIL"

    HikeRequest:
      type: object
      required:
        - name
        - mountainName
        - startAt
      properties:
        name:
          type: string
          example: "한라산 정상 도전"
        mountainName:
          type: string
          example: "한라산"
        startAt:
          type: string
          format: date-time
          example: "2025-08-20T09:00:00"

    HikeResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "한라산 정상 도전"
        mountainName:
          type: string
          example: "한라산"
        status:
          type: string
          enum: [PLANNED, IN_PROGRESS, COMPLETED]
          example: "PLANNED"
        startAt:
          type: string
          format: date-time
          example: "2025-08-20T09:00:00"
        endAt:
          type: string
          format: date-time
          example: "2025-08-20T15:00:00"
        memberCount:
          type: integer
          example: 5

    MessageRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          example: "여기 경치가 정말 좋네요!"
        latitude:
          type: number
          format: double
          example: 37.5665
          description: 위치 기반 메시지인 경우 위도
        longitude:
          type: number
          format: double
          example: 126.9780
          description: 위치 기반 메시지인 경우 경도

    MessageResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        content:
          type: string
          example: "여기 경치가 정말 좋네요!"
        userId:
          type: integer
          example: 123
        userNickname:
          type: string
          example: "등산러버"
        location:
          $ref: '#/components/schemas/LocationResponse'
        createdAt:
          type: string
          format: date-time
          example: "2025-08-15T10:30:00"